Refactor DB to handle favorite and current books
controllers/bookRoutes.js contains template data on
how to use the new schema to extract this info. DON'T MESS WITH IT.
Will actually set up endpoints tomorrow.

Front-end should expect a 'User' object with a 'BookJoin' indicating
favorite/current status with associated bookId.
This can be flipped to get a 'Book' object with a 'BookJoin' that
holds a userId.